Add TextureSize test
This test exposes two issues:
1. On NVIDIA drivers, the HLSL Texture.GetDimensions instruction doesn't
implicitly add the Texture's base level to the lod parameter.
This is inconsistent with the behaviour of the reference rasterizer.

2. When the HLSL shader calls GetDimensions on a Texture whose backing SRV
only specifies a limited number of mips, NVIDIA drivers can query properties
of the whole texture, while the reference rasterizer returns default values
for mips that are out of range of the SRV.